+++ src/constants/mock/listing-card-dummies.ts
